requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
openpyxl==3.1.2
pandas==2.2.2
celery==5.4.0
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
greenlet==3.0.3
